+++ screens/Notused/GetNews.js
@@ -8,7 +8,7 @@ import {
   Alert,
 } from 'react-native';
 import { ScrollView, TouchableOpacity } from 'react-native-gesture-handler';
-import config from '../config';
+import config from '../../config';
 
 const deviceHeight = Dimensions.get('window').height;
 const deviceWidth = Dimensions.get('window').width;
